# openapi.json
{
  "components": {
    "schemas": {
...
      "ChurnRiskAccount": {
        "additionalProperties": false,
        "properties": {
          "account_id": {
            "title": "Account Id",
            "type": "string"
          },
          "account_name": {
            "title": "Account Name",
            "type": "string"
          },
          "health_score": {
            "title": "Health Score",
            "type": "number"
          },
          "main_risk_driver": {
            "title": "Main Risk Driver",
            "type": "string"
          },
          "mrr": {
            "title": "Mrr",
            "type": "number"
          },
          "plan_tier": {
            "title": "Plan Tier",
            "type": "string"
          },
          "priority_rank": {
            "title": "Priority Rank",
            "type": "integer"
          },
          "recommended_action": {
            "title": "Recommended Action",
            "type": "string"
          },
          "region": {
            "title": "Region",
            "type": "string"
          },
          "risk_level": {
            "title": "Risk Level",
            "type": "string"
          },
          "segment": {
            "title": "Segment",
            "type": "string"
          }
        },
        "required": [
          "account_id",
          "account_name",
          "segment",
          "region",
          "plan_tier",
          "mrr",
          "health_score",
          "risk_level",
          "main_risk_driver",
          "recommended_action",
          "priority_rank"
        ],
        "title": "ChurnRiskAccount",
        "type": "object"
      },
      "CustomerAccountDetail": {
        "additionalProperties": false,
        "properties": {
          "account_id": {
            "title": "Account Id",
            "type": "string"
          },
          "account_name": {
            "title": "Account Name",
            "type": "string"
          },
          "avg_active_days": {
            "title": "Avg Active Days",
            "type": "number"
          },
          "current_mrr": {
            "title": "Current Mrr",
            "type": "number"
          },
          "customer_success_manager": {
            "title": "Customer Success Manager",
            "type": "string"
          },
          "failed_payments": {
            "title": "Failed Payments",
            "type": "number"
          },
          "health_score": {
            "title": "Health Score",
            "type": "number"
          },
          "main_risk_driver": {
            "title": "Main Risk Driver",
            "type": "string"
          },
          "orders_processed": {
            "title": "Orders Processed",
            "type": "number"
          },
          "plan_tier": {
            "title": "Plan Tier",
            "type": "string"
          },
          "recommended_action": {
            "title": "Recommended Action",
            "type": "string"
          },
          "region": {
            "title": "Region",
            "type": "string"
          },
          "restaurant_type": {
            "title": "Restaurant Type",
            "type": "string"
          },
          "risk_level": {
            "title": "Risk Level",
            "type": "string"
          },
          "segment": {
            "title": "Segment",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "support_calls": {
            "title": "Support Calls",
            "type": "number"
          },
          "weighted_pipeline_amount": {
            "title": "Weighted Pipeline Amount",
            "type": "number"
          }
        },
        "required": [
          "account_id",
          "account_name",
          "restaurant_type",
          "region",
          "segment",
          "plan_tier",
          "status",
          "current_mrr",
          "health_score",
          "risk_level",
          "main_risk_driver",
          "recommended_action",
          "customer_success_manager",
          "avg_active_days",
          "orders_processed",
          "support_calls",
          "failed_payments",
          "weighted_pipeline_amount"
        ],
        "title": "CustomerAccountDetail",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "AWS Serverless Support Analytics",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/customer-analytics/accounts/{account_id}": {
      "get": {
        "operationId": "get_account_detail_api_customer_analytics_accounts__account_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "title": "Account Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerAccountDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Account Detail",
        "tags": [
          "customer-analytics"
        ]
      }
    },
...
    "/api/customer-analytics/churn-risk": {
      "get": {
        "operationId": "get_churn_risk_api_customer_analytics_churn_risk_get",
        "parameters": [
          {
            "in": "query",
            "name": "risk_level",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Risk Level"
            }
          },
          {
            "in": "query",
            "name": "segment",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Segment"
            }
          },
          {
            "in": "query",
            "name": "region",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Region"
            }
          },
          {
            "in": "query",
            "name": "plan_tier",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Plan Tier"
            }
          },
          {
            "in": "query",
            "name": "minimum_mrr",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Minimum Mrr"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ChurnRiskAccount"
                  },
                  "title": "Response Get Churn Risk Api Customer Analytics Churn Risk Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Churn Risk",
        "tags": [
          "customer-analytics"
        ]
      }
    },
...
  }
}